requirements.txt
scikit-learn
groq
langgraph
langchain
langchain_groq
langchain-core
streamlit
pandas
pydantic
python-dotenv